DELIA-57608:gstaamp - g_object_set_valist: object class 'GstAppSrc' has no property named <logging string>
Reason for change: Typecasted the local variable as guint64 for setting "max-bytes" on g_object_set
Test Procedure: Check for warning - GstAppSrc has no property named
'Setting gst Audio buffer max bytes to %d'
Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_22_1

+++ aampgstplayer.cpp
@@ -990,14 +990,14 @@ static void InitializeSource(AAMPGstPlayer *_this, GObject *source, MediaType me
 		int MaxGstVideoBufBytes = 0;
 		_this->aamp->mConfig->GetConfigValue(eAAMPConfig_GstVideoBufBytes,MaxGstVideoBufBytes);
 		AAMPLOG_INFO("Setting gst Video buffer max bytes to %d", MaxGstVideoBufBytes);
-		g_object_set(source, "max-bytes", MaxGstVideoBufBytes, NULL);			/* Sets the maximum video buffer bytes as per configuration*/
+		g_object_set(source, "max-bytes", (guint64)MaxGstVideoBufBytes, NULL);			/* Sets the maximum video buffer bytes as per configuration*/
 	}
 	else if (eMEDIATYPE_AUDIO == mediaType || eMEDIATYPE_AUX_AUDIO == mediaType)
 	{
 		int MaxGstAudioBufBytes = 0;
                 _this->aamp->mConfig->GetConfigValue(eAAMPConfig_GstAudioBufBytes,MaxGstAudioBufBytes);
 		AAMPLOG_INFO("Setting gst Audio buffer max bytes to %d", MaxGstAudioBufBytes);
-		g_object_set(source, "max-bytes", MaxGstAudioBufBytes, NULL);			/* Sets the maximum audio buffer bytes as per configuration*/
+		g_object_set(source, "max-bytes", (guint64)MaxGstAudioBufBytes, NULL);			/* Sets the maximum audio buffer bytes as per configuration*/
 	}
 	g_object_set(source, "min-percent", 50, NULL);								/* Trigger the need data event when the queued bytes fall below 50% */
 	g_object_set(source, "format", GST_FORMAT_TIME, NULL);						/* "format" can be used to perform seek or query/conversion operation*/
